+++ packages/drivers/knex/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.5](https://github.com/proteinjs/db/compare/@proteinjs/db-driver-knex@1.13.4...@proteinjs/db-driver-knex@1.13.5) (2026-09-03)
+
+**Note:** Version bump only for package @proteinjs/db-driver-knex
+
+
+
+
+
+## [1.13.4](https://github.com/proteinjs/db/compare/@proteinjs/db-driver-knex@1.13.3...@proteinjs/db-driver-knex@1.13.4) (2026-09-02)
+
+**Note:** Version bump only for package @proteinjs/db-driver-knex
+
+
+
+
+
 ## [1.13.3](https://github.com/proteinjs/db/compare/@proteinjs/db-driver-knex@1.13.2...@proteinjs/db-driver-knex@1.13.3) (2026-09-02)
 
 **Note:** Version bump only for package @proteinjs/db-driver-knex