+++ packages/file-storage-drivers/google-cloud-storage/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.21](https://github.com/proteinjs/db/compare/@proteinjs/db-file-storage-driver-gcs@1.4.20...@proteinjs/db-file-storage-driver-gcs@1.4.21) (2026-09-03)
+
+**Note:** Version bump only for package @proteinjs/db-file-storage-driver-gcs
+
+
+
+
+
+## [1.4.20](https://github.com/proteinjs/db/compare/@proteinjs/db-file-storage-driver-gcs@1.4.19...@proteinjs/db-file-storage-driver-gcs@1.4.20) (2026-09-02)
+
+**Note:** Version bump only for package @proteinjs/db-file-storage-driver-gcs
+
+
+
+
+
 ## [1.4.19](https://github.com/proteinjs/db/compare/@proteinjs/db-file-storage-driver-gcs@1.4.18...@proteinjs/db-file-storage-driver-gcs@1.4.19) (2026-09-02)
 
 **Note:** Version bump only for package @proteinjs/db-file-storage-driver-gcs