+++ packages/settings/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.32](https://github.com/brentbahry/db/compare/@proteinjs/db-settings@1.3.31...@proteinjs/db-settings@1.3.32) (2026-09-03)
+
+**Note:** Version bump only for package @proteinjs/db-settings
+
+
+
+
+
+## [1.3.31](https://github.com/brentbahry/db/compare/@proteinjs/db-settings@1.3.30...@proteinjs/db-settings@1.3.31) (2026-09-02)
+
+**Note:** Version bump only for package @proteinjs/db-settings
+
+
+
+
+
 ## [1.3.30](https://github.com/brentbahry/db/compare/@proteinjs/db-settings@1.3.29...@proteinjs/db-settings@1.3.30) (2026-09-02)
 
 **Note:** Version bump only for package @proteinjs/db-settings